Bug fix and clean up
Fixed a bug that would result in the first sunrise failing if the app was installed after midnight and before sunrise.  Eliminated conditional branches that aren't required since the schedule() method ignores date information (was using if to index to the next day if rise or set was in the past).